bugfix add reservation

+++ src/main/java/at/qe/sepm/skeleton/ui/controllers/ReservationDetailController.java
@@ -8,14 +8,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
+import javax.faces.context.FacesContext;
+
 @Component
 @Scope("view")
 public class ReservationDetailController {
     @Autowired
     private EquipmentReservationService equipmentReservationService;
-
-    @Autowired
-    private EquipmentService equipmentService;
 
     /**
      * Attribute to cache the currently displayed reservation
@@ -43,7 +42,19 @@ public class ReservationDetailController {
      * Action to delete the currently displayed reservation.
      */
     public void doDeleteReservation() {
-        this.equipmentReservationService.deleteReservation(equipmentReservation);
-        equipmentReservation = null;
+        int idFromView = Integer.parseInt(
+            FacesContext
+                .getCurrentInstance()
+                .getExternalContext()
+                .getRequestParameterMap()
+                .get("detailReservation"));
+
+        this.setReservationId(idFromView);
+
+        if(equipmentReservation != null)
+        {
+            this.equipmentReservationService.deleteReservation(equipmentReservation);
+            equipmentReservation = null;
+        }
     }
 }
